Gut Meson Build Stuff and Use pyproject.toml Instead
Basically, this is just refactoring around pyproject.toml. Meson is gone
as a dependency, since this is all Python. The application script is
merged into megatech/vulkan/applications/DispatchTableGenerator.py
because pyproject.toml handles accessing it properly. The test file is
renamed to make it discoverable. Coverage is added as a test dependency.