WIP - Various Changes Towards 1.0
This commit is a big change. Basically, before I was emitting command
names apropos of nothing. This is fine for a dispatch table but, as my
use cases are expanding, I need to resolve the dependencies between
Vulkan features properly. To that end, I've added dependency
resolution to the library. Unfortunately, that also means the current
application is broken.

To enable proper resolution of dependencies, I've written a simple
tokenizer/lexer. In my opinion, it's pretty rudimentary but it gets the
job done. Vulkan dependencies only have 2 operators ("+" for the logical
conjuction and "," for the logical disjunction) anyway.

This application's new behavior is to leverage this dependency mechanism
for two cases. First, if a user tries to enable an extension without its
dependencies, the application will report an error. Second, extensions
can now be used to generate a C-style header guard that matches the
Vulkan dependency.

Here's a summary of the individual changes.

VulkanFeatures.py:

- Added Tokenizer classes and internal helper functions
- Added a VulkanRequirement class (this is exported) to represent each
  requirement
- Changed VulkanFeature so that commands are resolved directly to
  objects inside the constructor.
- Removed the VulkanFeature.commands method and replaced it with
  VulkanFeature.requirements
- Added VulkanFeature.dependency
- Added VulkanFeature.to_header_guard

VulkanCommand.py:

- Fixed some weird documentation
- Added VulkanCommandSet.find to allow name -> VulkanCommand mappings

VulkanSpecification.py

- Changed VulkanSpecification.apis to VulkanSpecification.features which
  is more consistent

DispatchTableGenerator.py:

- Ensured that the logger flushes after every message. I was getting
  weird output before this
- Added header guard to the verbose logging of a feature
- Added helper methods for finding enabled features
- Gutted DispatchTableGenerator.run and replaced it with a new approach
  that is dependency aware

TTFN